requirements.txt
python 
requests   
pandas          
numpy         
python-dotenv
pyyaml

#scraper engines
curl_cffi
beautifulsoup4 
lxml 

#warehouse
duckdb

#Just for testing
coloredlogs
